requirements.txt
# Requirements file placeholder
flask==2.3.3
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
tqdm==4.66.1
werkzeug==3.0.1
